+++ CHANGES.md
@@ -1,7 +1,7 @@
 0.8.0
 ==================
-
 * Moved from hard coded mappings to mapping files.
+* Used the line number colour for the print margin.
 
 0.7.0 (2011-01-29)
 ==================